add photo support

+++ src/controllers/photo/photo.controller.ts
@@ -1,10 +1,12 @@
 import { Request, Response, NextFunction, query } from 'express';
-import { JsonController, Post, Get, BodyParam, Body, QueryParams, Req, QueryParam, Param, Patch, Delete, Authorized, CurrentUser, MethodNotAllowedError, InternalServerError, UploadedFile } from 'routing-controllers';
+import { JsonController, Post, Get, BodyParam, Body, QueryParams, Req, QueryParam, Param, Patch, Delete, Authorized, CurrentUser, MethodNotAllowedError, InternalServerError, UploadedFile, Controller, NotFoundError, Res, NotAcceptableError } from 'routing-controllers';
 
 import * as fse from 'fs-extra';
 import * as path from 'path';
 import { PhotoDto } from './dto/photo.dto';
 import { PhotoModel } from '@packages/mongoose';
+import * as util from 'util';
+import { config_get } from '@packages/core';
 
 // export const fileUploadOptions = () => {
 //     storage: multer.diskStorage({
@@ -21,20 +23,28 @@ import { PhotoModel } from '@packages/mongoose';
 //     }
 // };
 
+export const PHOTO_BASE_PATH = config_get("photo.path","./dist/upload") ;
+
 @JsonController('/photo')
 export class PhotoController{
     @Authorized()
     @Post()
-    async upload(@UploadedFile("photo") file: any, @Body() dto:PhotoDto) {
+    async upload(@UploadedFile("photo") file: any, @Body({required:false}) dto:PhotoDto) {
         dto.album = dto.album || 'default' ;
-        dto.name = dto.name || file.filename ; 
+        dto.name = dto.name || file.originalname ; 
         dto.type = dto.type || file.mimetype ;
         // const obj =  Object.assign({name:file.filename,type:file.mimetype,album:'default'},dto) ;
+
+        let existPhoto = await PhotoModel.findOne({name:dto.name,album:dto.album}).exec() ;
+
+        if(existPhoto){
+            throw new NotAcceptableError("photo exists")
+        }
 
         const photo = new PhotoModel(dto) ;
 
         if(file.buffer) {
-            const filePath = path.join("./dist/upload",photo.path()) ;
+            const filePath = path.join(PHOTO_BASE_PATH,photo.path()) ;
             await fse.ensureFile(filePath) ;
             await fse.writeFile(filePath,file.buffer) ;
         }
@@ -43,5 +53,35 @@ export class PhotoController{
 
         return photo ;
     }
+
+}
+
+
+@Controller('/image')
+export class ImageController{
+
+    @Get('/:id([0-9a-f]{24})')
+    async getImage(@Param('id') id:string, @Res() res: Response) {
+        const photo = await PhotoModel.findById(id) ;
+        
+        if(photo) {
+            const filePath = path.join(__dirname,'../../../', PHOTO_BASE_PATH,photo.path()) ;
+            
+            console.log('File: ', filePath);
+            
+            // res.sendFile(filePath,);            
+
+            await util.promisify<void, void>(()=>{
+                res.sendFile(filePath)
+            })();
+            
+            return res ;
+            // throw null ;
+            
+        }
+        
+        return null ;
+    }
     
+
 }
